Consolidate finishedKeyguardStep and make it a StateFlow
Reroute the usages of finishedKeyguardStep to the flexiglass-aware
variants like isFinishedIn. This makes sure that future devs don't
missuse this API.

There is one unintended change in this refactor, where I have noticed
that for all isFinished... methods we explicitly call
distinctUntilChanged() while finishedKeyguardStep is not that. We should
be consistent here instead across the API. In this case we also want to
move to StateFlows which are always distinctUntilChanged which means
that I have now made it consistent in being distinct. If there is
actually any feature that relies on that we would need to keep it a
sharedFlow and remove the distinct call on all methods instead.

Bug: b/330311871
Flag: EXEMPT refactor
Test: REFACTOR_ONLY
Change-Id: I07be68ab6b0ce07c82c006e69ea0ddbcd076fb07