Implement resizable tiles prototype
Tapping a tile in edit mode no longer adds or removes it. Instead, it selects the tile that was tapped.
A selected tile will show a colored border and resizing handle to the right.
Dragging that handle will resize the tile between 1x1 and 2x1.

Bug: 350984160
Test: manually
Test: ResizingStateTest
Test: MutableSelectionStateTest
Test: ResizingTest
Flag: com.android.systemui.qs_ui_refactor

Change-Id: Ic4cb3fbb0348aa53fa9e49f251046f151f92d9f1

Former-commit-id: 25e03a2cd434f87c6bcbe0e57dc490204a7819ca