[sat] Always show icon if Connected or On
Currently the behavior of the device-based satelilte icon only relied on
all connections being OOS. There are 2 main issues with this:

1. If the radio entered the On or Connected states, we would still wait
   the 10s hysteresis time before showing the connected icon
2. If, for any reason, we got a Connected state when the OOS condition
   was not met, we would fail to show the icon

This change rearranges the logic a bit so that we have the same
provisioned / allowed check, and breaks out the hysteresis visibility
state from the connected/on visibility state. The flows are overall a
bit simpler.

Also added more table logs so we can track more clearly what is going
on.

Bug: 359356088
Flag: NONE bugfix
Test: DeviceBasedSatelliteViewModelTest
Test: manual using satellite demo mode
Change-Id: I4a0d94a29d713b050bab96e8322c577361d3668f