Fix FINISHED/CANCELED step filtered out incorrectly
The terminal step was filtered out because the transitioning condition
in STL is not met anymore as STL just moves on to any other
ObservalableTransitionState without guarantee of order. We need to look
at the previous transition to understand if the terminal step belongs
to one or another.

Bug: 361785511
Flag: com.android.systemui.scene_container
Test: tests modified to match real transition sequence of KTF/STL
Change-Id: I5cdd8be86c48903a5ad0d8025d9b93e9891f7b0f