Don't set exit animation state in DreamOverlayAnimationsController.wakeUp
wakeUp no longer actually controls the exit animation and thus we
shouldn't set the state to exit animation running state to true. This
caused the DreamOverlayService resetCurrentDreamOverlayLocked to never
finish when unlocking from the dream when showing the bouncer, which
doesn't run the exit animation.

Bug: 361864211
Fix: 361864211
Test: atest DreamOverlayAnimationsControllerTest DreamOverlayServiceTest
Flag: EXEMPT simple bugfix
Change-Id: Iba90f3dd3d369e6a58b6d1a086fae0b51e8e3b92