Skip falsing on non-touchscreen device sources
Previously we were analyzing the motion event classification to
determine if the device was an accurate tool (stylus, touchpad, etc)
in order to skip falsing.

Instead, we should perform the inverse, and can make the determination
that if the device source does not support SOURCE_TOUCHSCREEN, then
falsing should not be run as there is minimal risk of accidental touches

Additionally, we should not delay processing of MotionEvents on these
types of input.

Bug: 319809270
Test: manual - tested on tablet with trackpad attached
Test: manual - tested on tablet with mouse attached
Test: atest FalsingDataProviderTest.java
Flag: com.android.systemui.non_touchscreen_devices_bypass_falsing
Change-Id: Ie73e5da63452d2b5e15ed830c0a94f03718137a5

Former-commit-id: e1c367ebf2c67863502c19265940856a623cbf24