Expose new STLState.seekFoo() APIs
This CL exposes new APIs on MutableSTLState to easily seek a transition.
The APIs are based on the predictive back APIs: the transitions are
driven by a Flow and the transition is confirmed once the flow stops
emitting normally, or cancelled if the flow is cancelled.

Given that those new APIs are suspend functions, waiting for the
animation to finish or cancelling it works out of the box.

Bug: 362727477
Test: atest SceneTransitionLayoutStateTest
Flag: com.android.systemui.scene_container
Change-Id: Idc36c1eb2fff7c5f299e36cc2a495b921b878cf1

Former-commit-id: 86128b7f56c44f2989a4d7753f93a3a5f5b89413