Don't retry starting the artd_pre_reboot service.
`waitForService` uses an infinite loop to wait for the service. If the
service cannot start or register itself, the method will loop forever.
This is not suitable for artd_pre_reboot because it's a service from the
new system image, which is expected to fail if there is any
incompatibility.

This CL replaces the `waitForService` call with a `ctl.start` command
and a 5s wait.

Bug: 356858364
Test: adb shell pm art pr-dexopt-job --run
Change-Id: Id48f1c8f86c3c44e0a5524c35c91c470902b9468