Send latest configuration for binding process
Before a process completes attaching, its configuration won't be
updated with system. Then if there is a global change happens in
this period, the client process will get old configuration which
is assigned when creating the process record.

So this change popluates the needed info in a single invocation
when the process is attaching. Then the client can start with the
latest configuration.

This would also benefit:
- Reduce entering WM lock from AM's compatibilityInfoForPackage.
- Copy configuration inside WM lock to ensure data consistency.

Bug: 362215030
Flag: EXEMPT bugfix
Test: atest ActivityTaskManagerServiceTests

Change-Id: I4af30c62864eefb9897120ba46ebd0c7e3fc8231