Hide IME switch button for single IME and subtype
Previously in [1] we started showing the IME switcher button also if we
had a single IME with a single subtype, if the IME specified a
(language) settings activity, to allow users to easily open this and add
keyboard languages. However, this is impacting users that accidentally
open the menu, as they previously did not have the button visible.

This reverts the change to also show the button if the current IME
specified a (language) settings activity.

  [1]: I1005cb6b10682f3e91a7ed847c290fcc206b5faa

Flag: android.view.inputmethod.ime_switcher_revamp
Test: show IME with a single IME and subtype enabled
Bug: 383059282
Change-Id: I590740e5e9dd8fca74000c69afc557e8a82836f8

Former-commit-id: a393b9ba1552d1dfd01d981d44878dd72ee1490e