Rename enable_pip2_implementation to enable_pip2
This is to migrate the pip2 read-only flag to a read-write one. We
renamed the flag to avoid causing confusion to the system.

The pip2 flag is not used early in the boot process, it's safe for us to
do this convert.

Updated also the tracking bug in aconfig

Flag: com.android.wm.shell.enable_pip2
Bug: 362824417
Bug: 311462191
Test: Flip the flag in Flag Flipper app
Change-Id: I19e93cc8f463d93bf1ebe2f7d63b77d46832ce00

Former-commit-id: 47ad23e2cf9618ce130be7547d38302ffef02cf1