Add support for NFC associated role services
This allows NFC APDU services to opt-in to sharing the same priority as
the role owner if the containing package and the role owner have the
same package signature. The is useful for apps that have these types of
services spread amongst multiple packages.

Bug: 366243361
Test: atest RegisteredAidCacheTest
Flag: android.nfc.nfc_associated_role_services

Change-Id: Id1ff0a7dbef12283a135508d43f66148c08fbba3

Former-commit-id: 9f6ecc9c9d163f7e0aea57de8fa768a71e5291f8