Abort AE saved state restoration if it is too late
Abort the AE saved state restoration if it is not yet able
to be restored before any Activity#onCreate to be called.

And start the Task root activity as a fresh restart as a
fallback approach. Note that this may not always be correct
(for example, an new Activity was expected to be started
on the reused Task, but is now showing the Task root activity),
but it should be a minimum fallback in case the app is not
following the best practice to set up the embedding rules
in Application#onCreate.

Bug: 289875940
Test: verified on demo app
Flag: com.android.window.flags.ae_back_stack_restore
Change-Id: If7d33ec66a2476adf24892e54a4856db9b3d020a

Former-commit-id: dc6e0ff9d5ad5fa1be6b4ec37891793dd448d387