Make EmptyShadeView a LaunchableView
This is necessary because origin transitions using the Animation
library's ActivityTransitionAnimator (used internally by the
NotificationActivityStarter) require the view being animated from to be
a LaunchableView to avoid visibility-related flicker.

Since the onClick behavior of the empty shade text is exactly the same
as the footer's manage/history button, we should launch the intent in
the same way for consistency.

Bug: 366003631
Bug: 308591475
Flag: android.app.modes_ui_empty_shade
Flag: EXEMPT unflaggable inheritance change
Test: manually tested that pressing the No notifications text works
correctly when flag is on

Change-Id: I7e8ce55855dd7d5f4599dd4c6dd9b6ce951b9da9

Former-commit-id: 2768fbe71d56b9a05bc00bb21bfac5e190ad9e64