Require hold-to-drag for App Handle drags
Adds a holding period functionality to DragDetector, which requires a
hold within the slop region to be maintained for X amount of ms before
ACTION_MOVEs outside the slop are allowed (reported to the event
handler). This functionality is enabled for the App Handle's drag
detector behind a flag, and disable for every other drag detector
(header, resize listener).

Also modifies e2e test to check the type of input before entering
desktop with drag, and simulates a hold-to-drag when the input is from a
touchscreen.

Flag: com.android.window.flags.enable_hold_to_drag_app_handle
Bug: 356409496
Test: atest WMShellUnitTests; atest PlatformScenarioTests
Change-Id: Ib57be0ce8b63aaa17ecc57b70d1629ab88c69787

Former-commit-id: 66112fa6c9d41d54c98cb61dc93fe20127e4c088