[Unfold transition] Ensure that we start fold animation after half-folded state
Currently, lastHingeAngleBeforeTransition could be stuck
in '0' degrees or some low angle if user rotates the device
after unfolding or if stopped in half-folded state.

If then user unfolded the device further and started folding
the device, the fold animation won't kick-in before we reach
this angle.

Added updating of this angle when we unfold past it
after a threshold.

Bug: 365706412
Test: atest DeviceFoldStateProviderTest
Flag: EXEMPT bugfix
Change-Id: I9e706d74fcb021b8381ca7ef5b43afa00aeaebf9

Former-commit-id: a7d50cbb22396d40ace4843eebe5d0db827c30f9