Only launchTaskBehind dream if dream is on top
We don't specifically care about whether the device is dreaming or not
as a dream doesn't necessarily have an activity. We should check if
moveToTop worked or not - and in case it didn't, if that was because
there was an always-on-top activity in front.

Using the callback sent from DreamManagerService to ActivityTaskManager
is unreliable because it will not update state in sync with activities
starting/stopping and leads to race conditions.

Flag: com.android.window.flags.remove_activity_starter_dream_callback
Test: atest DreamManagerServiceTests
Bug: 366452352
Change-Id: I620072b756ccd892931561c6c7c4236c16c8a20e

Former-commit-id: b09b7d2e768146cf8bff4441e3b6536049b2f0bf