Move Edit mode to be top level in QS
That way, it doesn't need a GridAnchor. Even if we added a GridAnchor to
EditMode, the problem is that AnimatedContent ends up composing both at
the same time when swapping and STL crashes because there are two
elements with the same name.

Test: manual
Bug: 353254353
Flag: com.android.systemui.qs_ui_refactor_compose_fragment

Change-Id: I71735183f28221b69e4220991b6865b2ae8fbffb

Former-commit-id: 8e5da82a77e96f68b60d1365a168d25ce9ec5347