[flexiglass] Adds gestureFilter lambda to STL
System UI needs a way to be able to filter out drag gestures that start
(e.g. have their first down event) in specific regions on the display.
Please see the next CL on this chain to learn more about why.

This CL adds gestureFilter as a lambda that can be passed to STL and
logic where STL checks with gestureFilter to see whether it should
continue to handle a drag, given its starting position.

Bug: 367447743
Test: manually verified with the next CL
Test: compose tests added
Flag: NONE unused change, next CL is flagged
Change-Id: I081e7628f49681eb4a6d5630e91a0332944736e6

Former-commit-id: f9a01533ef8a416db85097c80e5ad7e89d0b4fd3