Stop app switches only when home is the focused app
The home activity was transient launched to top when transition started.
A translucent activity was started before the transition finished.
In that case, the app switch was stopped when finishing the transition
because the transition contains visible transient launches, while
the top activity was not home. Therefore, the top translucent activity
was unable to start activity and had to wait until timed out.

Bug: 368512259
Test: wm presubmit
Flag: EXEMPT bugfix
Change-Id: Ib46d76923f3bbacd06d6d9dea3f0bfc148b8c129

Former-commit-id: 3740c2e7b372c2978f5034186e5e2dea57137a97