[4/n] Add an API to disable edge resizing for tiling.
This change updates DesktopModeWindowDecoration with an new API,
updateDisabledResizingEdge, that disables a resizing edge including its
corners for finger, mouse and styles gestures.

This is needed because on tiling a task on the left, its right resizing
edge should be disabled, and on tiling a task on the right, its left
resizing edge should be disabled.

Flag: com.android.window.flags.enable_tile_resizing
Test: tests are WIP
Bug: 369043190
Change-Id: I06298eb08ef3f798bf11bd1823b7a2e113a39128

Former-commit-id: 830d7ccf7d09949a29857c10283bd12b20a7a924