Restore QSPanelControllerBase listening state
In flexiglass, when the views are detached we set the
tile layout to be not listening. However, if we were listening before,
we should still be listening after re-attaching. This makes sure that
that is restored properly.

Note that this applies to both QQS and QS (as they both use the same
abstract controller class).

Fixes: 366362115
Test: atest QSPanelControllerBaseTest
Test: atest PlatformScenarioTest:AirplaneModeIconTest
Flag: EXEMPT bugfix
Change-Id: Ia67ab43b67bf7699eeaca7687d6a9a70078e01d3

Former-commit-id: 79c8c820ad1e7fbd3eb1bf6d06ca3b44d40cd832