Find pending AppearedActivity first
Bug: 369460497

When the new activity is launched to the topmost TF because the source activity was in that TF, and the source activity is finished before resolving the new activity, we will try to see if the new activity match a rule with the split activities below. If matched, it can be reparented.

Change-Id: Ib5016fc5648f3f6d89b501e8d43014c915eb51e8
Signed-off-by: tanxiaoyan <tanxiaoyan@xiaomi.com>

Former-commit-id: 517947c85bf9506468b89b4ee402c57e911266fe