Fix Launching Foreground Recording App with MediaProjection App Selector
Partner found a bug relating to when we attempt to partial screenshare the app that is requesting MediaProjection. We always attempt to find the selected app by launch cookie provided, but if we select the currently open app that is requesting a MediaProjection, the launch cookie won't match. It only works now by accident because the launch cookie is attributed to the app selector activity instead of the base recording activity, but if the app selector activity is closed by the time virtual display set up starts then it won't find an activity with the new launch cookie. To remedy this, for the case of starting a mediaprojection for the active activity, then we'll provide the task id as a backup

Bug: 365498349
Test: atest ShellRecentTaskListProviderTest
Flag: EXEMPT bugfix
Change-Id: I7f5e4139dd72f7f7795adab00a52dc82b14e4ac9