Report audio-coupled haptics in Vibrator.isVibrating
Update VibratorController to set the isVibrating flag to true when the
vibrator enters external control. Starting from that point the audio
framework is free to send data to the vibrator motor directly, so the
service should reflect this state.

This will affect both the Vibrator.isVibrating and
OnVibratorStateChangedListener APIs.

Fix: 348589406
Flag: EXEMPT small bug fix
Test: VibratorControllerTest
Change-Id: I376c6c5773a3d5303bebe0a6e1c48121bf4e8ce8

Former-commit-id: e303f8e38e06140c77f35d9d95c8a4bec7725e04