Delete super-obsolete zen onboarding
A notification was shown when upgrading from pre-P to P because default DND settings had changed. The default was later changed back so it was already obsolete at that point. Moreover, it should be impossible to upgrade from pre-P to V in one step.

Bug: 346789065
Test: atest SystemNotificationChannelsTest for channel removal, otherwise N/A since this is code deletion
Flag: EXEMPT Unflaggable -- removing resources, manifest entries, etc.
Change-Id: I1c6af0e25b15108554c5fd57abff23da5d2a749e