Do not restart touch listening if TouchMonitor is destroyed.
This changelist ensures that TouchMonitor does not continue listening
and handling touches past destroy. Previously, this was possible if
touch handling ended after the monitor was destroyed. In this case, the
TouchMonitor was reset incorrectly to a listening/ready state.

Logging has also been improved with this change by adding a unique
identifier to each TouchMonitor instance for better attribution.

Test: atest TouchMonitorTest#testSessionPopAfterDestroy
Flag: EXEMPT bugfix
Fixes: 367549056
Change-Id: I16d99aab763570693144f15d102ae07ac2d9cd07

Former-commit-id: 831010e556becd2d9447f0851eb9b7ced9597bfa