Make App Header visibility follow status bar in full immersive
When the app is fully immersive, the app header should show and hide in
alignment with the status bar.
This change makes the caption visibility configurable with a
RelayoutParam, and disables the force-show of the caption when in
freeform, and instead checks whether the task is in full immersive mode
first.

Flag: com.android.window.flags.enable_fully_immersive_in_desktop
Bug: 369444183
Bug: 369444147
Bug: 369443876
Test: atest WMShellUnitTests
Test: check App Header follows status bar visibility
Change-Id: I7f39961f550ddf8656738e5f32a839b1c454bcad

Former-commit-id: f2c815d6d37657e7bdd613120e8162b8fff75f5a