Make shutdown text visible when wallpaper is white
The shutdown background was black. The wallpaper was white but hidden.
The shutdown text was trying to match wallpaper, so black was chosen,
which caused the text to be invisible.

This CL solves that issue by setting text color to
global_actions_shutdown_ui_text. Also this CL removes the code to blur
the background because its alpha has been set to 1 for a while now.

Bug: 193831592
Flag: EXEMPT resource only update
Test: atest ShutdownUiTest

Change-Id: I375beec37c26ce534cb4a5b9752acf5e40082aac

Former-commit-id: 955da7239e6ec26b86a6e8469641a1931e4b99b0