Revert MANAGE_KEY_GESTURES flag type to feature flag
Originally the flag was a feature flag, but I changed it to
bug fix but that is affecting the ability to ramp up the flag
in gantry. I need to either create a new flag and make changes
everywhere the flag is currently used or try to make this a
feature flag and ttry again in gantry.

Test: None
Bug: 361567988
Flag: com.android.hardware.input.manage_key_gestures
Change-Id: Ic9cc62c5f8b67aa7e15a438774bfef4966b03f10

Former-commit-id: a15238458fdc9941f610f4533043fe789f42a4b7