Disable ModesEmptyShadeFix in tests
This fixes the tests that are failing when modes_ui_empty_shade is
turned on.

For the view model tests, the new visibility should act the same as the
old one, so I kept the tests and I'm just getting the value from the
correct flow depending on the flag.
For the NSSL tests, they can just be disabled as we have equivalent
tests in the EmptyShadeViewModelTest.

Bug: 369859363
Bug: 366003631
Flag: TEST_ONLY
Test: presubmit with flag in staging

Change-Id: I66234cc1a6da9088fec19ee55b7081c56b947fc7

Former-commit-id: 45f85e804cbdf28854dfdd0a2b5ab489a609cc07