Opening type changes need alpha=1 when PiPing
In general, recents transitions opening tasks
or bringing them TO_FRONT have alpha=0 set
in Transitions#setupStartState().

So when we are entering PiP as a part of TRANSIT_OPEN,
for example, we need to reset the alpha of to_front/open
mode changes to 1f.

Bug: 359922654
Flag: com.android.wm.shell.enable_pip2
Test: tap home in 3-btn-nav from PiP-able activity
Change-Id: Id7cd8d65056e68a639b4ab38b0d2e55995bc72d5