SoundTriggerService.startRecognition - callingUser
SoundTriggerService.startRecognitionForService() uses the calling user
for setting up its RemoteSoundTriggerDetectionService, to inform the
caller of sound trigger events.

However, this method is first clearing the calling id, so when it
fetches the caller's userHandle, it will always get user 0, rather than
the actual caller.

This cl fixes that.

Bug: 341193639
Test: Manual - have an app in user 10 use this service and verify it receives the triggers
Flag: EXEMPT bugfix
Change-Id: I5bcec1797f65d53ff6110241ddd827c2dc4bf894

Former-commit-id: 8e423d11569e18a6c50c4b33bade4ee5f9a4a422