Create ManagerRecord#notifySessionUpdated
This change is part of a series of non-functional refactors to
encapsulate binder operations in the record that corresponds to the
target binder object. The objective is to improve the logging associated
with binder failures by adding information that's available in the
record. For example: Package names and ids.

Bug: 360129098
Test: atest CtsMediaBetterTogetherTestCases MediaRouter2HostSideTest
Flag: EXEMPT refactor
Change-Id: I9a91fa8bafaceed50ededf9f81b6c1db22412e5e

Former-commit-id: 33b673c15a5370660f5a33ad92be8c5597de5c04