Use UserHandle.isSameApp to determine if the uids are the same app
If we directly use uids to determine if they're the same app, it would
fail, due to the first digits of the uid is the userId, which is not
always 0.

Bug: 327711184
Test: Use device as a second user, open Settings > Accessibility. Use
adb to install an Accessibility app. Verify the newly installed
Accessibility app appears on the screen without leaving and coming back
to the screen.
Test: atest PackageMonitorCallbackHelperTest
Flag: EXEMPT low risk bugfix

Change-Id: Idf7043924f1c4228f7b3bf1e0d7c9cdfa9d3bdc7