Report legacyStart for all displays legacyPending reported
We send pending for every display that *might* have a change
on it; however, once the WM transaction is done (transition
ready), some of those changes might be no-ops. Clients of
the legacy callbacks expect pending to be followed by start,
so this change will report start for every display that
set pending even if it was ultimately no-op.

Bug: 356987205
Test: go to -1, then press back.
Flag: EXEMPT bugfix
Change-Id: I1adcb47d6a394a9b7fff3330d651b018e7910692

Former-commit-id: f2c5c50081449bc4688847ae7b5a6458b431ed3f