Revert "[flexiglass] Adds gestureFilter lambda to STL"
This reverts commit 22375717fad8b2a8d3406146b4a89b5d2bc15316 [formerly f9a01533ef8a416db85097c80e5ad7e89d0b4fd3].

Reason for revert: As discussed offline, the system gesture exclusion
was introduced for back gesture, so it should not be used for our
custom use cases. Moreover the gestureFilter API introduced in this CL
probably does not work as we would expect given that we still consume
(but ignore) all pointer events if the gesture is filtered. A better
name for this would have been "consumeButIgnoreGesture: (Offset) ->
Boolean", which highlights why this API should probably not be
introduced in the first place.

Change-Id: I2501b39e0a18b8659578eaac443a12a7a3da5102

Former-commit-id: 2ea0587736e8f857a5793d7b24f1e37d716479f3