Introduce Bounceable and Modifier.bounceable()
This CL introduces a new Modifier to make it easy to implement buttons
that can bounce, affecting the size of its neighbours. This will be used
initially by QuickSettings tiles and Magic Portrait effects.

See b/371173476#comment2 for a video and details about the
implementation.

Bug: 371173476
Test: atest BounceableTest
Flag: EXEMPT library code not used in feature code yet
Change-Id: Id376b49f05b400c3c492c26f1014ac31c1236343

Former-commit-id: 1ed596546e6dd28d967947f656e0864cefe33944