Content Recorder check mirror confirmation
In case DisplayManager boots longer than
WindowManager, the ContentRecorder may start
recording the content of the primary display
before the boot completes and without user
confirmation.

This is a privacy problem, so we need to wait
until user had a chance to confirm the mirroring.
This is happening after boot completes and the
display is enabled. This is because if user
not confirmed the mirroring, this can be
either during boot, or after boot when the display
is disabled. This is because the way for the user to
re-enable the display is either via confirmation
dialog or via settings, both ensures the
concious action by the user to enable mirroring.

Change-Id: I605d6d5cce8ecdf5d32a79d7e945fce0b59a29c9
Test: atest ExternalDisplayPolicyTest ContentRecorderTests
Bug: 361698995
Flag: com.android.server.display.feature.flags.enable_waiting_confirmation_before_mirroring

Former-commit-id: 63fe08d35430ec5ddf5d0d90bf229a1d569f9a64