Make userHandle non-null
Downstream code was already assuming it wasn't null. Not setting the
value in ScreenshotRequest would send in user zero which would not
result in a null. UserHandle.of() does not return null, so we'll always
have a value even if a bad int is passed in.

Bug: 370798587
Flag: EXEMPT minor refactor
Test: atest com.android.systemui.screenshot
Change-Id: Idbe4185b4bdafff090758b7d5faf276e39fa3184

Former-commit-id: 54dcbabd1bad5e9b6c1a08fdb7de5a3fb1056a89