Enable interruptions in Communal
This CL enables the enableInterruptions flag in the STLState used by
Communal. This flag is already enabled by default in STLState and is
going to be removed soon.

This CL shouldn't have any effect on Communal given that Communal has
only 2 scenes.

Bug: 371951148
Test: Manual, swiped from lockscreen to communal, entered edit mode,
 went back to communal then swiped back to lockscreen.
Flag: NONE this flag should not affect communal given that it has only 2
 scenes.
Change-Id: Ie3d0f2c09847d3d63f959a6007fd8f2c8c0da81c

Former-commit-id: d53f4c79bc3cf031494762a479ce7edcec0ba2dc