Initialise  visible_task_count state field of DesktopModeTaskUpdate atom.
Statsd needs an initial value for the state field of this atom,
otherwise metrics that slice by the state will be bucketed in an
UNKNOWN bucket.

To not intefere with any other uses of this atom, we define a new
task_event just to mark the initialisation.

Bug: 370478775
Flag: EXEMPT logging only
Test: Manual + atest DesktopModeLoggerTransitionObserverTest
Change-Id: I9efeef6d8a7264818ec0aac77368cb34fae53fa2