Reconcile MaterialTheme.colorScheme and AndroidColorScheme (1/2)
This CL reconcicles MaterialTheme.colorScheme with AndroidColorScheme,
so that the only colors left in AndroidColorScheme are the ones not
already available in MaterialTheme.colorScheme. See
b/370422346#comment2 for details.

This CL also adds a test that ensures that the Material colors match
the associated value coming from attributes (R.attr.materialColorFoo).

Bug: 370422346
Test: atest PlatformThemeTest
Flag: NONE color alignment
Change-Id: Ieef5db28a4b60e9de0d998597ad489c7d5468ebf

Former-commit-id: 6d09b190fd0fde5586a9b4862c174279780e975b