Ensure listeners are alerted to notif after LE
When a notification is posted, checks both the old _and_ new
notifications for the lifetime extension flag. If both have the flag,
then we know that this is the sysui-specific update occuring on a direct
reply. In this case, we don't need to alert any other listeners about
the update.

Previous versions of this code would only check the old notification for
the flag. On an update after a direct reply, the old notification would
have the flag. This would lead to the prepareNotifyPostedLocked function
hitting the "break;" statement as soon as system UI was alerted about
the notification, which meant that other listeners (like Wear) would not
be alerted.

Bug: 352471135
Flag: android.app.lifetime_extension_refactor
Test: atest NotificationListenersTest, flash and test
Change-Id: I036601a66ee94baa60e36217a1c392aee4bed5ab

Former-commit-id: cd8d4b7ac5e55533dc8d6eec7980321ef99a762d