PointerLocationView: Draw pointers in unrotated display space
The view is useful in debugging and recording where a user interacted on
screen. Previously, the view worked in the logical display space, like
any other app. This is not helpful in debugging input during rotation,
since the drawn trace content will rotate with the display.

In this CL, make PointerLocationView work in the unrotated display
space. In particular, the pointer trace (aka trail) and the pointer
states (e.g. crosshairs, velocity vector, touch/tool oval, etc.) are
drawn in the unrotated display space, while the display labels and text
will rotate with the display.

The view will now record where the user interacted with the device
relative the physical hardware, so the pointer trace will not rotate
with the display. This will give a more accurate representation of where
the user interactions happened when looking at screen recordings where
the display rotates.

Bug: 368317060
Test: manual, with setting enabled
Flag: EXEMPT bug fix
Change-Id: I4dfe797010914669c1aa8639fbb04e1eefd8ff13

Former-commit-id: 86494fef2b08c656aaf1aeabd235c5b2e71bb793