Report Activity as independent change when config-at-end
Config-at-end activities are actually changing independently
of their task. This means they should be collected and also
need to be reported even if they share changes with their
parent.

The upside of this is that WMCore no-longer needs to do shell
specific surface manipulations (pip pre-transforms) on the
activity surface since it is being sent to shell. Long-term
this also implies that we can remove some pip-specific logic
in core as well as some pip-specific communication.

Bug: 202201326
Test: atest TransitionTests#testConfigAtEnd
Flag: com.android.wm.shell.enable_pip2
Change-Id: I8800ce48e2b59c2f7928c8da3708bad712bedc5e