Remove wallpaper to not end up in empty desktop
In some cases, such as back navigation on a task with multiple task
fragments, we are not able to tell that the desktop will close after the
closing of the task on handleRequest. This cl adds a method to create
a transition remove desktop wallpaper if there is a single task left in
desktop and that's being removed in a transition.

Bug: 359899708
Test: atest DesktopTasksTransitionObserverTest
Test: Manually do back navigation on settings, observe that we don't end
up with empty desktop
Flag: EXEMPT Bugfix
Change-Id: Ia96d931a998be2665716f3aaf5fca7476c0aa602

Change-Id: I83c352b9aba7e2db895387a493925f29656d8828

Former-commit-id: e71724cbfd1e0d3690ef998ee3ae2d0bf0ee3eb2