Ignore dismiss requests if we're already going away.
These requests can come in if the activity we're unlocking to calls dismissKeyguard in onCreate/onResume. Activities that expect to sometimes launch over the lockscreen may do this (such as NewsActivity from Assistant). This should be a no-op, since we're already dismissing, but this was resulting in some activity launch animation state being set and never cleared, ending up on a black screen.

Test: manual with app that dismisses onResume
Fixes: 361832875
Flag: EXEMPT bugfix
Change-Id: Ibfc23643fe03a0f98fc69b2dea37e612b82d89ce

Former-commit-id: 655874ff0d41b7308ae3d62cf5e096730ca28e4c