Call dream overlay callback onWakeUp in dream overlay reset
The dream overlay may not always receive onWakeUp, which can cause
SysUI state to incorrect and not recognize that the user dream has
exited. This can happen when unlocking from the dream when the wake
redirect is active.

Always attempt to send onWakeUp when the overlay state is reset so that
SysUI state will be correct.

Bug: 370610704
Fixed: 370610704
Test: atest DreamOverlayCallbackControllerTest DreamOverlayServiceTest
Flag: android.service.dreams.dream_wake_redirect
Change-Id: I23716e7c884e41758a093461ec29f487f500bcbb

Former-commit-id: 899bdd1c81e3e17f0f81b101123a852086c03ddb