Make Javadoc changes to TtsSpan related to new API
We can't submit Javadoc changes to existing APIs that reference
FlaggedAPIs. So we put these changes in a different CL, which we should
submit before SDK finalization but after the quarterly releases that
come between API bumps.

This CL adds references to durations and seconds in the Javadocs related
to TYPE_TIME.

Flag: DOCS_ONLY
Bug: 337103893
Bug: 372323279
Test: builds
Change-Id: I75e55864379c104c47bb7bb50b475e8b8b14a1db

Former-commit-id: abff2386df0fb9edb5a6b304a1ddb6deb91cbc52