Validate TimeZone in TextClock.setTimeZone
TimeZone.getTimeZone appears to perform less stringent checks
than tz.ToZoneId, which leads to a crash in the widget when we do eventually call toZoneId upon a time tick.

An example of such an invalid timezone is "GMT+20:00".

TextClock is a remotable view, so this leads to crashes in the Launcher process when it's part of a widget.

Bug: 372811751
Change-Id: I3f0bf19830639387926f8ab2e685a7dd0d991958
Test: atest TextClockTest
Flag: EXEMPT Localized bug fix with limited impact

Former-commit-id: 96b623828e48ce818844784266b91951ca9e3196