Resubmit of 'Allow loading of host runtime for Windows"
This makes the loading of host runtime dependent on a Java system
property instead of simply ignoring it for Windows. By default, this
will be loaded (keeping the current behaviour working for non windows platforms).

But setting use_base_native_hostruntime to false will allow the host runtime to be
skipped when libandroid_runtime is loaded.

Flag: NONE host-only change
Bug: 340885449
Test: atest RavenwoodCoreTest
Change-Id: Idaf33dda2c114732ea4862c8ec203907b8e93042

Former-commit-id: c1f1411122038070d3fbab37442686526d5ede35