Fix ViewRootImpl traversal issue after SurfaceSyncGroup timeout
ViewRootImpl traversal is blocked from the time a sync is requested
to when the sync is successfully added to a SurfaceSyncGroup.
If the ViewRootImpl draw timeouts, the ViewRootImpl's
SurfaceSyncGroup is marked as ready. If there is another
sync request on the ViewRootImpl before the traversal path,
we can get in a state where we will never draw again.

Fix this by updating cleaning up the ViewRootImpl state
when the SurfaceSyncGroup is marked as ready.

Flag: EXEMPT bug fix
Bug: 362513091
Test: presubmit (coming in future cl)
Change-Id: I6c0ac8f96914d231cc9dc695f6fc8975ba2c5864

Former-commit-id: 89f28270932c1c42c42706285e99bb30e28ae4b6