Remove synchronization on getViewerString
This might have been causing lock contention in the main thread when called while waiting for protolog to be loaded or unloaded. It will be the callers responsibility to ensure the loading and unloading of configs has completed successfully before querying the viewer config if required.

Flag: EXEMPT fixing regression
Bug: 369833736
Test: atest TracingTests
Change-Id: Ica8203138c697897f63df30f145b14125554aeb3

Former-commit-id: 43103941a8840c503496e84531a1a96b7f50bb75