Add InterfaceDiedCallback to RemoteCallbackList
Some of the existing uses of RemoteCallbackList extend the class to
override onCallbackDied. However with the Builder pattern it is no
longer possible to do so.

This change offers an alternative that works with the Builder pattern. A
new Builder API setInterfaceDiedCallback() is added to allow the caller
to provide custom death handling logic.

Flag: android.os.binder_frozen_state_change_callback
Change-Id: Ib777d9920d8332a09a6da3601ad3bab071fb9cba
Test: atest --instant android.os.cts.RemoteCallbackListTest
Bug: 361157077
API-Coverage-Bug: 372923111

Former-commit-id: 5a0bc1ef9f697dbc973480b5aba4c07d420711a7