Improved screenshot policy for desktop and split mode
The new unified policy class simplifies the rules and behavior
of screenshots, handling situations with mixed profile content
on screen.

Improvements:

When two work profile tasks apear as split-screen mode, the
screenshot will now capture both.

When desktop mode is active with freeform windows, content
will be saved to the private profile if any is private profile,
otherwise to the personal profile.

When in desktop mode with a single maximized window, it is
treated as a full screen task and handled accordingly.

The focused task is no longer used to determine the owner of the
screenshot making the result less surprising to end users.

If work content appears beside personal content in split mode,
the screenshot will remain full-screen and be saved to the personal
profile.

Bug: 365597999
Flag: com.android.systemui.screenshot_policy_split_and_desktop_mode
Test: atest ScreenshotPolicyTest
Change-Id: I5353b477d527b79f8b95803d7ba6136fdba441be

Former-commit-id: d39fd18d5dd98bffe1419e9bf94c1d1c0eb64a06